+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.8.38] - 2025-09-10
+
+### Changed
+
+- Cleaner installer output without warnings
+
+### Fixed
+
+- Fix installer deprecation warning
+- Better compatibility across Python versions
+
+### Removed
+
+- Remove pkg_resources deprecation warning during installation
 ## [4.8.37] - 2025-09-10
 
 ### Changed